refactor: Refactored folders
