+++ src/App.jsx
@@ -2,22 +2,22 @@ import "./index.css";
 import { useState, createContext, useEffect, useMemo } from "react";
 import { Route, Routes, Navigate } from "react-router";
 import API from "./API/API";
-import DefaultLayout from "./components/DefaultLayout";
-import Login from "./components/Login";
-import Registration from "./components/Registration";
-import AdminPage from "./components/AdminPage";
-import MapPage from "./components/MapPage";
-import NotAuthorized from "./components/NotAuthorized";
+import DefaultLayout from "./components/Layout/DefaultLayout";
+import Login from "./components/Auth/Login";
+import Registration from "./components/Auth/Registration";
+import AdminPage from "./components/Admin/AdminPage";
+import MapPage from "./components/Map/MapPage";
+import NotAuthorized from "./components/Routes/NotAuthorized";
 import Homepage from "./components/Homepage";
 import HowItWorks from "./components/HowItWorks";
-import PublicRelationsOfficer from "./components/PublicRelationsOfficer";
+import PublicRelationsOfficer from "./components/InternalUsers/PublicRelationsOfficer";
 import LoadingSpinner from "./components/LoadingSpinner";
-import ProtectedRoute from "./components/ProtectedRoute";
-import RoleBasedRedirect from "./components/RoleBasedRedirect";
-import OfficerPage from "./components/OfficerPage";
+import ProtectedRoute from "./components/Routes/ProtectedRoute";
+import RoleBasedRedirect from "./components/Routes/RoleBasedRedirect";
+import OfficerPage from "./components/InternalUsers/OfficerPage";
 import { allowedOfficerRoles } from "./constants/allowedOfficerRoles";
-import UserPage from "./components/UserPage";
-import MaintainerPage from "./components/MaintainerPage";
+import UserPage from "./components/Citizen/UserPage";
+import MaintainerPage from "./components/InternalUsers/MaintainerPage";
 import WebSocketTest from "./components/WebSocketTest";
 
 export const NavbarTextContext = createContext();
