+++ src/components/Layout/Footer.jsx
@@ -1,0 +1,11 @@
+import "../styles/Footer.css";
+
+function Footer() {
+  return (
+    <footer className="footer-custom text-center py-3 mt-4">
+      <span>© 2025 PARTICIPIUM. All rights reserved.</span>
+    </footer>
+  );
+}
+
+export default Footer;
