+++ src/components/Map/MapLogic.jsx
